Move FUSE-specific code into a submodule
